+++ Source/JavaScriptCore/wasm/WasmMemory.h
@@ -90,7 +90,7 @@ public:
     Expected<PageCount, GrowFailReason> grow(VM&, PageCount);
     bool fill(uint32_t, uint8_t, uint32_t);
     bool copy(uint32_t, uint32_t, uint32_t);
-    bool init(uint32_t, const uint8_t*, uint32_t);
+    bool init(uint64_t, const uint8_t*, uint32_t);
 
     void registerInstance(JSWebAssemblyInstance&);
